Implement actual operations
At first I was going to keep addition, subtraction, division, and
multiplication as separate PR's/commits, but once I abstracted out the
Operation class they became so trivial as to not feel worth it.

---

Most of this has been added to the readme but here's a quick breakdown:

- Operation is a class that describes an interface that all of our
current mathematical operations implement. You hand them an array of
operands and they do their best to compute the output. One operand will
be computed with an `identity` operand. Zero operands is an error. Our
calculator basically ignores operations with no operands, so this
shouldn't come up, but I'm leaving the possible exception there because
I think it could be helpfully descriptive if something ever goes wrong
with a change to `Calculator`.
- Reading through the actual operations should be super obvious at this
point.

Other miscellaneous changes:

- The default gemfile has some non-rubocop conforming stuff so I edited
  that to quiet down rubocop.
- Without a target version rubocop won't acknowledge the
  frozen_string_literal pragma and kept complaining about mutable
  strings as constants
- I forgot to pass arguments through to the `rspec` binstub :whoops:

Next Steps:

- Now is probably a good time to implement an actual CLI interface to
play with this thing