Remove last command functionality
I actually really dislike this now that I've played with it for a
second. I feel like it *could* be cool, but I'd have to change the way
I handle output to handle the case where I want to apply the previous
command with only one item on the stack, and then start over with an
empty stack. Otherwise it's kind of just confusing.